[Refactor] 라네즈 제품 조회 API 공통 응답 설정 및 구조 수정

+++ src/main/java/com/pocketmon/insightpocket/domain/laneige/dto/LaneigeProductItem.java
@@ -8,8 +8,8 @@ public record LaneigeProductItem(
         String product_name,
         String style,
         BigDecimal price,
-        Integer rank_1,
+        Long rank_1,
         String rank_1_category,
-        Integer rank_2,
+        Long rank_2,
         String rank_2_category
 ) {}
